mirror/requirements.txt
opencv-python
mediapipe
tflite-runtime
runwayml
python-dotenv
requests
pydbus
